org-users move shared fn into test helpers
