+++ src/components/org-users/test-helpers.ts
@@ -1,0 +1,37 @@
+export function composeOrgRoles(setup: object) {
+  const defaultRoles = {
+    billing_managers: {
+      current: '0',
+    },
+    managers: {
+      current: '0',
+    },
+    auditors: {
+      current: '0',
+    },
+  };
+
+  return {
+    ...defaultRoles,
+    ...setup,
+  };
+}
+
+export function composeSpaceRoles(setup: object) {
+  const defaultRoles = {
+    developers: {
+      current: '0',
+    },
+    managers: {
+      current: '0',
+    },
+    auditors: {
+      current: '0',
+    },
+  };
+
+  return {
+    ...defaultRoles,
+    ...setup,
+  };
+}
